+++ example/maze/model/grid.dart
@@ -3,16 +3,13 @@ part of maze;
 //www.migapro.com/depth-first-search
 class MazeGrid extends Grid {
   MazeGrid(Table table): super(table) {
-    cellPieces.forEach((CellPiece cp) {
-      cp.tag.number = 1;
-      cp.color.main = 'black'; // wall
-    });
+    cellPieces.forEach((CellPiece cp) => cp.color.main = 'black'); // wall
     CellPiece sp = cellPieces.randomCellPiece();
     sp.tag.number = 0;
-    gen(sp.cell.column, sp.cell.row);
+    path(sp.cell.column, sp.cell.row);
   }
   
-  gen(int c, int r) {
+  path(int c, int r) {
     var rds = new List<Direction>();
     for (var i = 0; i < 4; i++) {
       rds.add(randomDirection());
@@ -26,10 +23,10 @@ class MazeGrid extends Grid {
           }
           var cp1 = cellPieces.cellPiece(c - 1, r);
           var cp2 = cellPieces.cellPiece(c - 2, r);
-          if (cp2.tag.number == 1) { // if wall
-              cp2.tag.number = 0;  // path
-              cp1.tag.number = 0;
-              gen(c - 2, r);
+          if (!cp2.tag.isMarked) {     // if wall
+              cp2.tag.isMarked = true; // path
+              cp1.tag.isMarked = true;
+              path(c - 2, r);
           }
           break;
         case Direction.RIGHT:
@@ -39,10 +36,10 @@ class MazeGrid extends Grid {
           }
           var cp1 = cellPieces.cellPiece(c + 1, r);
           var cp2 = cellPieces.cellPiece(c + 2, r);
-          if (cp2.tag.number == 1) {
-              cp2.tag.number = 0;
-              cp1.tag.number = 0;
-              gen(c + 2, r);
+          if (!cp2.tag.isMarked) {
+              cp2.tag.isMarked = true;
+              cp1.tag.isMarked = true;
+              path(c + 2, r);
           }          
           break;
         case Direction.UP:
@@ -52,10 +49,10 @@ class MazeGrid extends Grid {
           }
           var cp1 = cellPieces.cellPiece(c, r - 1);
           var cp2 = cellPieces.cellPiece(c, r - 2);
-          if (cp2.tag.number == 1) {
-              cp2.tag.number = 0;
-              cp1.tag.number = 0;
-              gen(c, r - 2);
+          if (!cp2.tag.isMarked) {
+              cp2.tag.isMarked = true;
+              cp1.tag.isMarked = true;
+              path(c, r - 2);
           }
           break;
         case Direction.DOWN:
@@ -65,10 +62,10 @@ class MazeGrid extends Grid {
           }
           var cp1 = cellPieces.cellPiece(c, r + 1);
           var cp2 = cellPieces.cellPiece(c, r + 2);
-          if (cp2.tag.number == 1) {
-              cp2.tag.number = 0;
-              cp1.tag.number = 0;
-              gen(c, r + 2);
+          if (!cp2.tag.isMarked) {
+              cp2.tag.isMarked = true;
+              cp1.tag.isMarked = true;
+              path(c, r + 2);
           }    
       }
     }