+++ example/maze/view/board.dart
@@ -3,7 +3,7 @@ part of maze;
 class Board extends Surface {
   Board(CanvasElement canvas, MazeGrid grid) : super(canvas, grid: grid) {
     for (var cp in grid.cellPieces) {
-      if (cp.tag.number == 0) {
+      if (cp.tag.isMarked) {
         cp.color.main = 'orange'; // path
       } 
     }